+++ mybatis-plus-core/src/main/java/com/baomidou/mybatisplus/core/toolkit/sql/segments/NormalSegment.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2011-2020, hubin (jobob@qq.com).
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.baomidou.mybatisplus.core.toolkit.sql.segments;
+
+import java.util.ArrayList;
+import java.util.stream.Collectors;
+
+import com.baomidou.mybatisplus.core.conditions.ISqlSegment;
+
+/**
+ * @author miemie
+ * @since 2018-06-27
+ */
+public class NormalSegment extends ArrayList<ISqlSegment> implements ISqlSegment {
+
+    private static final long serialVersionUID = -1991374407733611565L;
+
+    @Override
+    public String getSqlSegment() {
+        return this.stream().map(ISqlSegment::getSqlSegment).collect(Collectors.joining("and"));
+    }
+}
